requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
aiosqlite==0.17.0
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
ansible==9.10.0
ansible-core==2.16.11
ansible-runner==2.4.1
anyio==4.5.2
aodhclient==3.6.0
APScheduler==3.11.0
asgiref==3.8.1
astroid==3.3.9
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.2.0
autopage==0.5.2
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1;python_version<"3.9"
bcrypt==4.2.1
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.1.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cliff==4.7.0
cmd2==2.5.7
configobj==5.0.6
croniter==5.0.1
cryptography==43.0.3
dasbus==1.4
databases==0.9.0
debtcollector==3.0.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
distro==1.5.0
dnspython==2.6.1
docker==7.1.0
docopt==0.6.2
dogpile.cache==1.3.3
ecdsa==0.19.0
et_xmlfile==2.0.0
eventlet==0.38.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
fasteners==0.19
fastjsonschema==2.21.1
file-magic==0.4.0
flower==2.0.1
futurist==3.0.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
humanize==4.12.1
idna==2.10
immutables==0.21
#importlib_metadata==8.5.0
#importlib_resources==6.4.5
iniconfig==2.0.0
ipython==8.12.3
iso8601==2.1.0
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.21
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keystoneauth1==5.8.0
kombu==5.4.2
kubernetes==33.1.0
libcomps==0.1.20.post1
lockfile==0.12.2
loguru==0.5.3
logutils==0.3.5
lxml==5.3.0
Mako==1.3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.0.2
msgpack==1.1.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
ncclient==0.6.16
netaddr==1.3.0
netifaces==0.10.6
networkx==3.1
neutron-lib==3.16.0
numpy==1.24.4
oauthlib>=3.2.2
openpyxl==3.1.5
openstacksdk==4.0.1
os-client-config==2.1.0
os-ken==2.11.2
os-service-types==1.7.0
os-traits==3.2.0
osc-lib==3.2.0
oslo.cache==3.8.0
oslo.concurrency==6.1.0
oslo.config==9.6.0
oslo.context==5.6.0
oslo.db==16.0.0
oslo.i18n==6.4.0
oslo.log==6.1.1
oslo.messaging==14.9.0
oslo.metrics==0.9.0
oslo.middleware==6.2.0
oslo.policy==4.4.0
oslo.reports==3.4.0
oslo.serialization==5.5.0
oslo.service==3.5.0
oslo.upgradecheck==2.4.0
oslo.utils==7.3.0
oslo.versionedobjects==3.4.0
osprofiler==4.2.0
ovs==3.4.1
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
Paste==3.10.1
PasteDeploy==3.1.0
pbr==6.1.0
pecan==1.5.1
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.2
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
prettytable==0.7.2
python-swiftclient==4.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.6.0
pure_eval==0.2.3
pyasn1==0.6.1
pycadf==4.0.0
pycparser==2.22
pydantic==1.10.0
pydantic_core==2.27.2
pydeps==3.0.1
pydot==3.0.2
Pygments==2.18.0
PyJWT==2.9.0
pylint==3.3.6
pymemcache==4.0.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyperclip==1.9.0
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
pytest==8.3.5
python-daemon==3.1.2
python-barbicanclient==7.0.0
python-blazarclient==4.1.0
python-cinderclient==9.6.0
python-dateutil==2.9.0.post0
python-designateclient==6.1.0
python-glanceclient==4.7.0
python-heatclient==3.5.0
python-ironicclient==5.8.0
python-jose==3.3.0
python-keystoneclient==5.5.0
python-mistralclient==5.2.0
python-monascaclient==2.7.0
python-multipart==0.0.17
python-neutronclient==11.3.1
python-novaclient==18.6.0
python-octaviaclient==3.8.0
python-openstackclient==7.1.3
python-troveclient==8.6.0
python-vitrageclient==5.1.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
repoze.lru==0.7
requests==2.32.3
requestsexceptions==1.4.0
resolvelib==1.0.1
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.20.1
rsa==4.9
setproctitle==1.3.4
six==1.15.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.40.0
statsd==4.0.1
stdlib-list==0.11.1
stevedore==5.3.0
tenacity==9.0.0
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.2
tinycss2==1.4.0
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.5
uvicorn==0.17.6
vine==5.1.0
warlock==2.0.1
wcwidth==0.2.13
webencodings==0.5.1
WebOb==1.8.9
websocket-client==1.8.0
wrapt==1.17.0
yappi==1.6.10
yaql==3.0.0
yarg==0.1.9
zipp==3.20.2
websockets==13.1
redis==5.2.1
dbutils==3.1.0
reportlab==4.4.1
